Fix counting bug
This commit addresses the proximitychanged event being thrown
without user input by only checking for the condition where
the last sensor state was NOT the the maximum range. According to
the docs the output of different platforms may be different, but
this approach should be more stable than keeping track of the
sensor's previous state with a flag as there doesn't seem to be
a way to properly prevent the proximitychanged event from being
thrown when a user switches between a non-proximity enabled app
and this one.